# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Registro de Pontos",
        "description": "API para gerenciar pontos dos funcionários",
        "contact": {
            "email": "diegohsananias@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/points": {
            "get": {
                "tags": [
                    "Points"
                ],
                "summary": "Listar pontos",
                "description": "Listar os pontos do usuário autenticado ou de um subordinado (se for administrador).",
                "operationId": "listPoints",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "ID do usuário para filtrar pontos (somente para administradores)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "start_date",
                        "in": "query",
                        "description": "Data inicial do filtro",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2024-11-01"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "description": "Data final do filtro",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2024-11-30"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de pontos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "datetime": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2024-11-26T10:30:00Z"
                                            },
                                            "latitude": {
                                                "type": "number",
                                                "format": "float",
                                                "example": "-23.550520"
                                            },
                                            "longitude": {
                                                "type": "number",
                                                "format": "float",
                                                "example": "-46.633308"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Dados inválidos"
                    }
                }
            },
            "post": {
                "tags": [
                    "Points"
                ],
                "summary": "Registrar um ponto",
                "description": "Registrar um ponto com localização do funcionário",
                "operationId": "registerPoint",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "datetime": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2024-11-26T10:30:00Z"
                                    },
                                    "latitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": "-23.550520"
                                    },
                                    "longitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": "-46.633308"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ponto registrado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "datetime": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-11-26T10:30:00Z"
                                        },
                                        "latitude": {
                                            "type": "number",
                                            "format": "float",
                                            "example": "-23.550520"
                                        },
                                        "longitude": {
                                            "type": "number",
                                            "format": "float",
                                            "example": "-46.633308"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Dados inválidos"
                    }
                }
            }
        },
        "/points/{id}": {
            "put": {
                "tags": [
                    "Points"
                ],
                "summary": "Atualizar um ponto existente",
                "description": "Atualizar um ponto existente com base no ID.",
                "operationId": "updatePoint",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do ponto a ser atualizado",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "datetime": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2024-11-26T10:30:00Z"
                                    },
                                    "latitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": "-23.550520"
                                    },
                                    "longitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": "-46.633308"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ponto atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "datetime": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-11-26T10:30:00Z"
                                        },
                                        "latitude": {
                                            "type": "number",
                                            "format": "float",
                                            "example": "-23.550520"
                                        },
                                        "longitude": {
                                            "type": "number",
                                            "format": "float",
                                            "example": "-46.633308"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Dados inválidos"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Points"
                ],
                "summary": "Remover um ponto",
                "description": "Remover um ponto existente com base no ID.",
                "operationId": "deletePoint",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do ponto a ser removido",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Ponto removido com sucesso"
                    },
                    "404": {
                        "description": "Ponto não encontrado"
                    }
                }
            }
        },
        "/points/report": {
            "get": {
                "tags": [
                    "Points"
                ],
                "summary": "Relatório de pontos dos funcionários",
                "description": "Relatório detalhado de pontos registrados, com filtros opcionais por data.",
                "operationId": "pointsReport",
                "parameters": [
                    {
                        "name": "start_date",
                        "in": "query",
                        "description": "Data inicial do filtro",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2024-11-01"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "description": "Data final do filtro",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2024-11-30"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Relatório de pontos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "point_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "employee_name": {
                                                "type": "string",
                                                "example": "Carlos Silva"
                                            },
                                            "employee_role": {
                                                "type": "string",
                                                "example": "employee"
                                            },
                                            "employee_age": {
                                                "type": "integer",
                                                "example": 34
                                            },
                                            "manager_name": {
                                                "type": "string",
                                                "example": "João Gerente"
                                            },
                                            "point_date": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2024-11-10T08:30:00Z"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}